+++ ffindex-java-core/src/test/java/org/rcsb/ffindex/impl/ReadOnlyFileBundleTest.java
@@ -4,21 +4,20 @@ import org.junit.jupiter.api.Test;
 import org.rcsb.ffindex.Conversions;
 import org.rcsb.ffindex.FileBundleIO;
 import org.rcsb.ffindex.ReadableFileBundle;
+import org.rcsb.ffindex.TestHelper;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
-import java.nio.file.Paths;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 class ReadOnlyFileBundleTest {
     @Test
     void whenReadingFromFileBundle_thenContentMatchesExpectation() throws IOException {
-        Path resourcePath = Paths.get("src/test/resources/");
-        Path dataPath = resourcePath.resolve("test.data");
-        Path indexPath = resourcePath.resolve("test.ffindex");
+        Path dataPath = TestHelper.createTempFile("test.data");
+        Path indexPath = TestHelper.createTempFile("test.ffindex");
 
         try (ReadableFileBundle fileBundle = FileBundleIO.openBundle(dataPath, indexPath).inReadOnlyMode()) {
             assertEquals("a", Conversions.toString(fileBundle.readFile("a")));